% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_relative_velocity.R
\name{add_relative_velocity}
\alias{add_relative_velocity}
\title{Add relative velocity columns to motility data}
\usage{
add_relative_velocity(summary_dat, condition, control_name, ph_value)
}
\arguments{
\item{ph_value}{}
}
\description{
Add relative velocity columns to motility data
}
