% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_cardiac_motility.R
\name{analyze_cardiac_motility}
\alias{analyze_cardiac_motility}
\title{Analyze cardiac motility}
\usage{
analyze_cardiac_motility(data)
}
\arguments{
\item{data}{}
}
\value{
a summary dataframe. Works the same as analyze_motility() but with Distance >=1 inclusion criteria
}
\description{
Analyze cardiac motility
}
