% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_mini_ensemble.R
\name{analyze_mini_ensemble}
\alias{analyze_mini_ensemble}
\title{Analyze Mini Ensemble Data - Finds Event Duration, Force, and Time Offs}
\usage{
analyze_mini_ensemble(data, mv2nm, detrend)
}
\arguments{
\item{data}{}
}
\description{
Analyze Mini Ensemble Data - Finds Event Duration, Force, and Time Offs
}
