% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_motility.R
\name{analyze_motility}
\alias{analyze_motility}
\title{Analyze motility data}
\usage{
analyze_motility(data)
}
\arguments{
\item{data}{Data needs to be a dataframe from WRMTRCK}
}
\value{
A summary of motility data including the average velocity and percent moving.
    average_velocity calculated off the mean velocity of the "AvgSpeed" from all filaments moving more than 4um.
    Percent moving calculated by dividing the number of filaments that moved (>4um) by the total number of filaments.
    NOTE* Use with dplyr "group_by" to get output for each variables.
}
\description{
Automates the analysis of motility data from the WRMTRCK output
}
\examples{
analyze_motility()

data <- data(motility_test)
analyzed_data <- data \%>\%
group_by("tm", "pH", "pCa", "video") \%>\%
analyze_motility()

}
