% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_plot.R
\name{bin_plot}
\alias{bin_plot}
\title{Histogram plot from binned velocity data}
\usage{
bin_plot(data)
}
\arguments{
\item{desired_bins}{}
}
\value{
a histogram from continous data that has been binned
}
\description{
Histogram plot from binned velocity data
}
\examples{
bin_plot(data_from_mass_import)
}
