% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c1v1.R
\name{c1v1}
\alias{c1v1}
\title{Dilution/Concentration Calculator}
\usage{
c1v1(c1, v1, c2, v2)
}
\arguments{
\item{c1}{initial concentration}

\item{v1}{initial volume}

\item{c2}{final concentration}

\item{v2}{final volume}
}
\value{
a number
}
\description{
Calculates unknown concentrations or volumes using (c1 * v1 = c2 * v2). Make sure the numbers you give are in the same units! This function will not know otherwise. Whichever parameter is set to "find", the function will solve for.
}
\examples{

c1v1(c1 = 100, v1 = 2, c2 = "find", v2 = 1000)

}
