% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conc_calc.R
\name{conc_calc}
\alias{conc_calc}
\title{Concentration Calculator (c1 * v1 = c2 * v2)}
\usage{
conc_calc(c1, v1, c2, v2)
}
\arguments{
\item{v2}{}
}
\description{
Concentration Calculator (c1 * v1 = c2 * v2)
}
\examples{

}
