% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detrend_trap.R
\name{detrend_trap}
\alias{detrend_trap}
\title{Drift Correction for Laser Trapping Data}
\usage{
detrend_trap(x)
}
\arguments{
\item{x}{}
}
\value{
A Vector
}
\description{
Drift Correction for Laser Trapping Data
}
