% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drc.michaelis_menten.R
\name{drc.michaelis_menten}
\alias{drc.michaelis_menten}
\title{Michaelis Menten self-starter (fct) for drm w/custom param names}
\usage{
drc.michaelis_menten()
}
\value{

}
\description{
Michaelis Menten self-starter (fct) for drm w/custom param names
}
