% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drc.xDP_release_xTP_binding.R
\name{drc.xDP_release_xTP_binding}
\alias{drc.xDP_release_xTP_binding}
\title{xDP-release, xTP-binding self-starter (fct) for drm}
\usage{
drc.xDP_release_xTP_binding()
}
\value{

}
\description{
xDP-release, xTP-binding self-starter (fct) for drm
}
