% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_downread.R
\name{drive_downread}
\alias{drive_downread}
\title{Download and read google drive file (read_sheets is sooo slow)}
\usage{
drive_downread(dribble, col_names, type)
}
\arguments{
\item{col_names}{}
}
\value{

}
\description{
Download and read google drive file (read_sheets is sooo slow)
}
