% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_hidden_markov_analysis_desktop.R,
%   R/drive_hidden_markov_analysis_shiny.R
\name{drive_hidden_markov_analysis_shiny}
\alias{drive_hidden_markov_analysis_shiny}
\title{HMM Analysis}
\usage{
drive_hidden_markov_analysis_shiny(date_dribble, mv2nm, nm2pn, overlay_color)

drive_hidden_markov_analysis_shiny(date_dribble, mv2nm, nm2pn, overlay_color)
}
\value{



}
\description{
HMM Analysis

HMM Analysis
}
\examples{
This is an interactive plot of the running variance and running mean.
The alogorithm for event detection is a Variance-Hidden-Markov-Model.
Shaded regions are the events identified by the model.




Below is an interactive plot of the raw data (model does not use this data).
The overlay is the Hidden Markov Model 'state' prediction (via Viterbi Algorithm) multipled by several conversion factors.
These convert the x-axis 'time' from windows to data points to seconds.
The HMM state, baseline mean, and measured step size are used to scale the overlay to each step and subsequent baseline.


Plots of the running mean and the running variance.
This is a useful way to see how the model decides what is an event or baseline.
This is an interactive plot of the running variance and running mean.
The alogorithm for event detection is a Variance-Hidden-Markov-Model.
Shaded regions are the events identified by the model.




Below is an interactive plot of the raw data (model does not use this data).
The overlay is the Hidden Markov Model 'state' prediction (via Viterbi Algorithm) multipled by several conversion factors.
These convert the x-axis 'time' from windows to data points to seconds.
The HMM state, baseline mean, and measured step size are used to scale the overlay to each step and subsequent baseline.


Plots of the running mean and the running variance.
This is a useful way to see how the model decides what is an event or baseline.
}
