% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_make_trap_observations_shiny.R,
%   R/simulation_drive_make_trap_observations_shiny.R
\name{drive_make_trap_observations_shiny}
\alias{drive_make_trap_observations_shiny}
\title{Make Trap Observations}
\usage{
drive_make_trap_observations_shiny(date_folder)

drive_make_trap_observations_shiny(date_folder)
}
\arguments{
\item{txts}{}
}
\value{



}
\description{
Make Trap Observations

Make Trap Observations
}
