% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_writeup.R
\name{drive_writeup}
\alias{drive_writeup}
\title{Write file to temp. directory and upload to google drive}
\usage{
drive_writeup(data, drive_location, name, type, overwrite, col_names)
}
\arguments{
\item{col_names}{}
}
\value{

}
\description{
Write file to temp. directory and upload to google drive
}
