% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dygraph_clean_shave.R
\name{dygraph_clean_shave}
\alias{dygraph_clean_shave}
\title{Creates a dygraph to clean trap data in shiny}
\usage{
dygraph_clean_shave(data, first_file_name, title)
}
\arguments{
\item{first_file_name}{}
}
\value{

}
\description{
Creates a dygraph to clean trap data in shiny
}
