% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dygraph_trim_and_shave.R
\name{dygraph_trim_and_shave}
\alias{dygraph_trim_and_shave}
\title{Creates a dygraph to clean trap data in shiny}
\usage{
dygraph_trim_and_shave(data, first_file_name)
}
\arguments{
\item{data}{}

\item{first_file_name}{}
}
\value{

}
\description{
Creates a dygraph to clean trap data in shiny
}
