% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble_average.R
\name{ensemble_average}
\alias{ensemble_average}
\title{Forward & Backward Ensemble Averages}
\usage{
ensemble_average(events)
}
\arguments{
\item{events}{}
}
\value{
Both forward and backward ensemble averages
}
\description{
Forward & Backward Ensemble Averages
}
