% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equipartition.R
\name{equipartition}
\alias{equipartition}
\title{Equipartition method for calibrating trap stiffness}
\usage{
equipartition(data)
}
\arguments{
\item{data}{}
}
\value{
trap stiffness
}
\description{
Equipartition method for calibrating trap stiffness
}
\examples{
equipartition(vector)
}
