% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_frequency.R
\name{event_frequency}
\alias{event_frequency}
\title{Calculate 'real-time' event frequency of laser trapping data}
\usage{
event_frequency(processed_data, rle_object, conversion)
}
\arguments{
\item{processed_data}{a numeric vector of processed laser trap data}

\item{rle_object}{a df of run_length encoding from HM-Model}

\item{conversion}{a number that converts between running window time and 5kHz}
}
\value{
a dataframe of event frequencies per each second
}
\description{
Calculate 'real-time' event frequency of laser trapping data
}
