% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_peaks.R
\name{find_peaks}
\alias{find_peaks}
\title{Find Peaks Function from "https://github.com/stas-g/findPeaks"}
\usage{
find_peaks(x, m = 3)
}
\arguments{
\item{m}{}
}
\value{

}
\description{
Find Peaks Function from "https://github.com/stas-g/findPeaks"
}
\examples{
The function takes an ordered sequence (vector) of values x and a number m and returns a vector of indices of local peaks in x. A (local) peak is defined as a point such that m points either side of it has a lower or equal value to it. Thus, m can be used adjust the sensitivity of the peak detection procedure: larger m will result in fewer peaks, whilst smaller values of m will result in more peaks found.
}
