% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_exp.R
\name{fit_exp}
\alias{fit_exp}
\title{Fit Exponential Distribution}
\usage{
fit_exp(data, bw, color, zoom)
}
\arguments{
\item{data}{}

\item{bw}{}

\item{color}{}

\item{zoom}{}
}
\value{

}
\description{
Fit Exponential Distribution
}
