% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_hill.R
\name{fit_hill}
\alias{fit_hill}
\title{Fits Hill Equation}
\usage{
fit_hill(data)
}
\arguments{
\item{data}{}
}
\value{
a drm object (list)
}
\description{
Fits Hill equation using drc::drm(). This function for a single condition in a dataframe. See fit_hill_nest() to fit multiple variables.
Data frame must have a "average_velocity" and "pCa" column. If not, use drc::drm() directly. This function simply calls the drm() function and provides some of the parameter inputs for quicker analysis.
}
\examples{

x <- data(motility_test)
fit_data <- fit_hill(x)

}
