% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_hill_nest_possible_pm.R
\name{fit_hill_nest_possible_pm}
\alias{fit_hill_nest_possible_pm}
\title{Fit Hill Equation to nested data}
\usage{
fit_hill_nest_possible_pm(data, nest_key)
}
\arguments{
\item{nest_name}{}
}
\description{
Fit Hill Equation to nested data
}
