% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_hill_pm.R
\name{fit_hill_pm}
\alias{fit_hill_pm}
\title{Fits Hill Equation to percent moving data}
\usage{
fit_hill_pm(data)
}
\arguments{
\item{data}{}
}
\value{
a drm object (list)
}
\description{
Fits Hill Equation to percent moving data
}
