% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_dropbox_path.R
\name{format_dropbox_path}
\alias{format_dropbox_path}
\title{Formats dropbox path for shiny upload}
\usage{
format_dropbox_path(x)
}
\arguments{
\item{x}{A file path returned by drop_dir()$path_lower.}
}
\value{

}
\description{
Formats dropbox path for shiny upload
}
