% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_coefficients.R
\name{get_coefficients}
\alias{get_coefficients}
\title{Make tibble of Hill Equation parameters from a fitted drm model}
\usage{
get_coefficients(drm_model, conditions)
}
\arguments{
\item{conditions}{}
}
\value{
a tibble of coefficients with an identifier column
}
\description{
Make tibble of Hill Equation parameters from a fitted drm model
}
