% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hidden_markov_analysis_shiny.R
\name{hidden_markov_analysis_shiny}
\alias{hidden_markov_analysis_shiny}
\title{HMM Analysis}
\usage{
hidden_markov_analysis_shiny(parent_dir, mv2nm, nm2pn, overlay_color)
}
\value{

}
\description{
HMM Analysis
}
