% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histo_plotly.R
\name{histo_plotly}
\alias{histo_plotly}
\title{Create Plotly Histograms}
\usage{
histo_plotly(wd, trap_analysis, colors)
}
\arguments{
\item{colors}{}
}
\value{

}
\description{
Create Plotly Histograms
}
\examples{

}
