% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_mini_events.R
\name{identify_mini_events}
\alias{identify_mini_events}
\title{Identifies mini ensemble events (>10nm, >10ms)}
\usage{
identify_mini_events(raw_data, run_mean)
}
\arguments{
\item{run_mean}{}
}
\description{
Identifies mini ensemble events (>10nm, >10ms)
}
