% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_processed_trap.R
\name{import_processed_trap}
\alias{import_processed_trap}
\title{Import processed trapping data files}
\usage{
import_processed_trap(working_dir)
}
\arguments{
\item{working_dir}{characters string of path to directory where files are located}
}
\value{
One data frame. Recommended to use tidyr::separate() on 'filenames[i]' columns.
}
\description{
Imports all files in the current working directory into list
}
\examples{
import_txt()


trap_data <- import_txt("my_wd")




}
