% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_shiny.R
\name{import_shiny}
\alias{import_shiny}
\title{Import Files to R Shiny}
\usage{
import_shiny(shiny_datapath, file_name)
}
\arguments{
\item{file_name}{}
}
\value{

}
\description{
Import Files to R Shiny
}
