% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_tsv.R
\name{import_tsv}
\alias{import_tsv}
\title{Import .tsv files}
\usage{
import_tsv(working_dir)
}
\arguments{
\item{working_dir}{characters string of path to directory where files are located}
}
\value{
One data frame. Recommended to use tidyr::separate() on 'filenames[i]' columns.
}
\description{
Imports and concactenates all .tsv files in the current working directory
}
\examples{
import_tsv()


motility_data <- import_tsv("my_wd") \%>\%
separate('filenames[i]', c("date, pH, mutation, pCa"), sep = "_")




}
