% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_xls.R
\name{import_xls}
\alias{import_xls}
\title{Import .xls files}
\usage{
import_xls(working_dir)
}
\arguments{
\item{working_dir}{characters string of path to directory where files are located}
}
\value{
One data frame. Recommended to use tidyr::separate() on 'filenames[i]' columns.
}
\description{
Imports and concactenates all files in the current working directory
}
\examples{
import_xls()


motility_data <- import_xls("my_wd") \%>\%
separate('filenames[i]', c("date, pH, mutation, pCa"), sep = "_")




}
