% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_trap_observations.R
\name{make_trap_observations}
\alias{make_trap_observations}
\title{Make Trap Observations}
\usage{
make_trap_observations(wd)
}
\arguments{
\item{txts}{}
}
\value{

}
\description{
Make Trap Observations
}
