% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_trap_observations_shiny.R
\name{make_trap_observations_shiny}
\alias{make_trap_observations_shiny}
\title{Make Trap Observations}
\usage{
make_trap_observations_shiny(wd)
}
\arguments{
\item{txts}{}
}
\description{
Make Trap Observations
}
