% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure_mini.R, R/measure_mini_durations.R
\name{measure_mini}
\alias{measure_mini}
\title{Measure ton and toff for mini ensemble data}
\usage{
measure_mini(raw_data, run_mean)

measure_mini(raw_data, run_mean)
}
\arguments{
\item{run_mean}{}

\item{run_mean}{}
}
\description{
Measure ton and toff for mini ensemble data

Measure ton and toff for mini ensemble data
}
