% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure_mini_ensemble.R
\name{measure_mini_rescaled}
\alias{measure_mini_rescaled}
\title{Measure ton and toff for mini ensemble data}
\usage{
measure_mini_rescaled(raw_data, run_mean)
}
\arguments{
\item{run_mean}{}
}
\value{

}
\description{
Measure ton and toff for mini ensemble data
}
