% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mini_ensemble_analyzer.R
\name{mini_ensemble_analyzer}
\alias{mini_ensemble_analyzer}
\title{Mini Ensemble Analysis}
\usage{
mini_ensemble_analyzer(parent_dir, mv2nm, nm2pn, run_mean_color)
}
\arguments{
\item{run_mean_color}{}
}
\value{

}
\description{
Mini Ensemble Analysis
}
