% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mini_ensemble_analyzer_gui.R
\name{mini_ensemble_analyzer_gui}
\alias{mini_ensemble_analyzer_gui}
\title{Mini Ensemble Analysis for GUI}
\usage{
mini_ensemble_analyzer_gui(parent_dir, mv2nm, nm2pn, run_mean_color)
}
\arguments{
\item{run_mean_color}{}
}
\value{

}
\description{
Mini Ensemble Analysis for GUI
}
