% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mini_ensemble_analyzer_shiny.R
\name{mini_ensemble_analyzer_shiny}
\alias{mini_ensemble_analyzer_shiny}
\title{Mini Ensemble analyzer for shiny}
\usage{
mini_ensemble_analyzer_shiny(input, detrend, baseline_start_sec,
  baseline_end_sec, mv2nm, nm2pn, run_mean_color)
}
\arguments{
\item{run_mean_color}{}
}
\description{
Mini Ensemble analyzer for shiny
}
