% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mini_find_true_end.R
\name{mini_find_true_end}
\alias{mini_find_true_end}
\title{Find first data point not in mini ensemble event}
\usage{
mini_find_true_end(x)
}
\arguments{
\item{x}{}
}
\description{
Find first data point not in mini ensemble event
}
