% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mini_rescale_start.R
\name{mini_rescale_start}
\alias{mini_rescale_start}
\title{Find better event start}
\usage{
mini_rescale_start(raw_data, run_mean)
}
\arguments{
\item{run_mean}{}
}
\value{

}
\description{
Find better event start
}
