% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_distr.R
\name{plot_distr}
\alias{plot_distr}
\title{Quick histogram plotting}
\usage{
plot_distr(data, aes_x, bw, xlab, colors)
}
\arguments{
\item{colors}{}
}
\value{
a ggplot
}
\description{
Quick histogram plotting
}
