% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_motility.R
\name{plot_motility}
\alias{plot_motility}
\title{Plot a single velocity-pCa curve}
\usage{
plot_motility(data, predict_data)
}
\arguments{
\item{predict_data}{}
}
\value{
a ggplot scatterplot fit with the hill equation
}
\description{
Data needs to be from analyze_motility() and predict_hill()
}
