% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_motility2.R
\name{plot_motility2}
\alias{plot_motility2}
\title{Plot motility data for 2 velocity-pCa curves}
\usage{
plot_motility2(data_7.4, predict_7.4, data_6.8, predict_6.8)
}
\arguments{
\item{predict_6.8}{}
}
\value{
a ggplot scatterplot fit with the hill equation
}
\description{
Plot motility data for 2 velocity-pCa curves
}
