% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/possibly_hill_nest.R
\name{possibly_hill_nest}
\alias{possibly_hill_nest}
\title{Determime if Hill Equation can be fit}
\usage{
possibly_hill_nest(data, nest_key)
}
\arguments{
\item{nest_name}{}
}
\description{
This will attempt to fit the Hill Equation to data in a nested data frame. If hill equation cannot be fit NA is reported. You can use dplyr::filter() to remove these and use fit_hill_nest() to remaining data.
}
