% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/possibly_hill_nest_pm.R
\name{possibly_hill_nest_pm}
\alias{possibly_hill_nest_pm}
\title{Fit Hill Equation to Nested Percent Moving Data}
\usage{
possibly_hill_nest_pm(data, nest_key)
}
\arguments{
\item{nest_name}{}
}
\description{
Fit Hill Equation to Nested Percent Moving Data
}
