% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_hill.R
\name{predict_hill}
\alias{predict_hill}
\title{Make Hill Equation predictions for plotting}
\usage{
predict_hill(drm_model)
}
\arguments{
\item{drm_model}{}
}
\value{
a tibble of theoretetical pCa values and their estimated y-values using the parameters generated by the fit of the drm model
}
\description{
Make Hill Equation predictions for plotting
}
\examples{
predict_hill(my_drm_model)
}
