% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_mm.R
\name{predict_mm}
\alias{predict_mm}
\title{Make MM Equation predictions for plotting}
\usage{
predict_mm(mm_model)
}
\arguments{
\item{mm_model}{}
}
\value{
a tibble of theoretetical conc values and their estimated y-values using the parameters generated by the fit of the MM model
}
\description{
Make MM Equation predictions for plotting
}
\examples{
predict_hill(my_drm_model)
}
