% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round_any.R
\name{round_any}
\alias{round_any}
\title{Round to any integer from https://stackoverflow.com/questions/43627679/round-any-equivalent-for-dplyr/46489816#46489816}
\usage{
round_any(x, accuracy, f = round)
}
\arguments{
\item{accuracy}{}
}
\value{

}
\description{
Round to any integer from https://stackoverflow.com/questions/43627679/round-any-equivalent-for-dplyr/46489816#46489816
}
