% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/running_mv.R
\name{running_mv}
\alias{running_mv}
\title{Calulate Running Mean and Variance}
\usage{
running_mv(data, window_width, by)
}
\arguments{
\item{window_width}{}
}
\description{
Calulate Running Mean and Variance
}
