% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_group.R
\name{sheets_group}
\alias{sheets_group}
\title{Group sheets on Google Drive}
\usage{
sheets_group(read_from, upload_to)
}
\arguments{
\item{upload_to}{}
}
\value{

}
\description{
Group sheets on Google Drive
}
