% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_clean_dygraph.R
\name{shiny_clean_dygraph}
\alias{shiny_clean_dygraph}
\title{Generate dygraph for cleaning trap data in shiny}
\usage{
shiny_clean_dygraph(trap_selected_obs, curr_clean_graph)
}
\arguments{
\item{trap_selected_obs}{}
}
\value{

}
\description{
Generate dygraph for cleaning trap data in shiny
}
