% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_make_trap_observations.R
\name{shiny_make_trap_observations}
\alias{shiny_make_trap_observations}
\title{Make Trap Observations}
\usage{
shiny_make_trap_observations(trap_selected_date, threshold)
}
\arguments{
\item{txts}{}
}
\value{

}
\description{
Make Trap Observations
}
