% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_mini_ensemble_analyzer.R
\name{shiny_mini_ensemble_analyzer}
\alias{shiny_mini_ensemble_analyzer}
\title{Mini-ensemble Analyzer}
\usage{
shiny_mini_ensemble_analyzer(
  trap_selected_date,
  trap_selected_conditions,
  mv2nm,
  nm2pn,
  color,
  file_type
)
}
\arguments{
\item{file_type}{}
}
\value{

}
\description{
Mini-ensemble Analyzer
}
