% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_move_trap.R
\name{shiny_move_trap}
\alias{shiny_move_trap}
\title{Move trap files to new observations}
\usage{
shiny_move_trap(
  trap_selected_date,
  trap_obs,
  trap_selected_obs,
  trap_files,
  dygraph_clean_date_window_1,
  dygraph_clean_date_window_2
)
}
\arguments{
\item{trap_files}{}
}
\value{

}
\description{
Move trap files to new observations
}
