% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_summarize_trap.R
\name{shiny_summarize_trap}
\alias{shiny_summarize_trap}
\title{Summarize trap project (shiny app)}
\usage{
shiny_summarize_trap(trap_selected_project, file_type)
}
\arguments{
\item{file_type}{}
}
\value{

}
\description{
Summarize trap project (shiny app)
}
