% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_trap_upload.R
\name{shiny_trap_upload}
\alias{shiny_trap_upload}
\title{Move trap files (.txt) to Box Sync folder}
\usage{
shiny_trap_upload(input_data, trap_selected_date)
}
\arguments{
\item{datapath}{}
}
\value{

}
\description{
Move trap files (.txt) to Box Sync folder
}
\examples{

}
