% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_trim_dygraph.R
\name{shiny_trim_dygraph}
\alias{shiny_trim_dygraph}
\title{Trim trap grouped4r file in shiny}
\usage{
shiny_trim_dygraph(
  trap_selected_obs,
  trap_grouped_file,
  input_dygraph_clean_shave_date_window_1,
  input_dygraph_clean_shave_date_window_2
)
}
\arguments{
\item{trap_selected_obs}{}
}
\value{

}
\description{
Trim trap grouped4r file in shiny
}
