% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_trap.R
\name{simulate_trap}
\alias{simulate_trap}
\title{Simulate Laser Trap Data}
\usage{
simulate_trap(number_of_events, signal_2_noise)
}
\arguments{
\item{signal_2_noise}{}
}
\value{
a vector
}
\description{
Simulate Laser Trap Data
}
\examples{

simulate_trap(10, 2)

}
