% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_trap_custom.R
\name{simulate_trap_custom}
\alias{simulate_trap_custom}
\title{Simulate Laser Trap Data}
\usage{
simulate_trap_custom(
  number_of_events,
  step_size,
  event_duration,
  signal_2_noise,
  add_noise
)
}
\arguments{
\item{signal_2_noise}{}
}
\value{
a vector
}
\description{
Simulate Laser Trap Data
}
\examples{

simulate_trap(10, 2)

}
