% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/squysh_time.R
\name{squysh_time}
\alias{squysh_time}
\title{Prints the system date/time in only numbers (no spaces or punctuation)}
\usage{
squysh_time()
}
\value{

}
\description{
Prints the system date/time in only numbers (no spaces or punctuation)
}
