% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_cal.R
\name{step_cal}
\alias{step_cal}
\title{Performs Step Calibration of the Laser Trap}
\usage{
step_cal(vector, step)
}
\arguments{
\item{vector}{}
}
\value{

}
\description{
Performs Step Calibration of the Laser Trap
}
