% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_mini_events.R
\name{summarize_mini_events}
\alias{summarize_mini_events}
\title{Summarize Mini Ensemble Event}
\usage{
summarize_mini_events(data)
}
\arguments{
\item{data}{}
}
\value{
a summary tibble
}
\description{
Summarize Mini Ensemble Event
}
