% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_trap_shiny.R
\name{summarize_trap_shiny}
\alias{summarize_trap_shiny}
\title{Summarize Trap Analysis for Shiny}
\usage{
summarize_trap_shiny(wd, trap_analysis)
}
\arguments{
\item{trap_analysis}{}
}
\value{
a .csv file to the wd name *_trap_summary.csv
}
\description{
Summarize Trap Analysis for Shiny
}
\examples{
summarize_trap_shiny("my/directory", "mini")



}
