% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_ppt.R
\name{theme_ppt}
\alias{theme_ppt}
\title{ggplot2 theme for powerpoint presentations}
\usage{
theme_ppt(base_size = 28, base_family = "")
}
\arguments{
\item{base_rect_size}{}
}
\value{

}
\description{
ggplot2 theme for powerpoint presentations
}
