% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_upload_trap_shiny.R
\name{upload_trap_shiny}
\alias{upload_trap_shiny}
\title{Upload trap files (.txt) to google drive as sheets}
\usage{
upload_trap_shiny(input_data, destination)
}
\arguments{
\item{datapath}{}
}
\value{

}
\description{
Upload trap files (.txt) to google drive as sheets
}
\examples{

}
