% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_temp_csv.R
\name{write_temp_csv}
\alias{write_temp_csv}
\title{Writes temporary (.csv) file}
\usage{
write_temp_csv(x, filename, ...)
}
\arguments{
\item{filename}{}
}
\value{
a character string
}
\description{
Writes temporary (.csv) file
}
\examples{
tempfile <- write_temp_csv(iris, "iris.csv")
}
