% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_hidden_markov_analysis.R
\name{shiny_hidden_markov_analysis}
\alias{shiny_hidden_markov_analysis}
\title{Hidden Markov for Shiny (dropbox)}
\usage{
shiny_hidden_markov_analysis(
  trap_selected_date,
  trap_selected_conditions,
  mv2nm,
  nm2pn,
  overlay_color,
  file_type,
  hm_emcontrol
)
}
\arguments{
\item{overlay_color}{}
}
\value{

}
\description{
Hidden Markov for Shiny (dropbox)
}
\examples{
This is an interactive plot of the running variance and running mean.
The algorithm for event detection is a Hidden Markov Model and these are the data the model receives as input. The model is fitted with the EM algorithm and the
gray shaded regions are the binding events identified through state sequence decoding via the Viterbi alogorithm.




Below is an interactive plot of the raw data (model does not use this data).
The overlay is the Hidden Markov Model 'state' prediction multipled by several conversion factors.
These convert the x-axis (time) from 'windows' to 'data points' to 'seconds'.
The HMM state, baseline mean, and measured step size are used to scale the model overlay to each step and subsequent baseline level.


Plots of the running mean vs. running variance.
This provides insight into how the model divided data into either the baseline or event populations.
}
