% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_mini_ensemble_analyzer.R,
%   R/shinymini_ensemble_analyzer.R
\name{shiny_mini_ensemble_analyzer}
\alias{shiny_mini_ensemble_analyzer}
\title{Mini Ensemble analyzer for shiny}
\usage{
shiny_mini_ensemble_analyzer(trap_selected_date, mv2nm, nm2pn, color)

shiny_mini_ensemble_analyzer(trap_selected_date, mv2nm, nm2pn, color)
}
\arguments{
\item{run_mean_color}{}
}
\value{



}
\description{
Mini Ensemble analyzer for shiny

Mini Ensemble analyzer for shiny
}
