% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_motility.R
\name{summarize_motility}
\alias{summarize_motility}
\title{Summarize output from 'analyze_motility'}
\usage{
summarize_motility(data)
}
\arguments{
\item{analyze_motility_data}{}
}
\description{
Summarizes the data from analyze_motility() that can be used for plotting average of all analyzed data.
}
